+++ fluidSimLib/timer.cpp
@@ -14,12 +14,12 @@ timer::~timer( )
     std::cout << mInfo << " took " << durationInSeconds() << " seconds" << std::endl;
 }
 
-double timer::durationInSeconds() const 
+FloatType timer::durationInSeconds() const 
 {
     auto end = std::chrono::steady_clock::now();
     auto count = std::chrono::duration_cast<std::chrono::microseconds>(
         end - mStartingPoint).count();
-    return count / 1000000.0;
+    return count / (FloatType)1000000.0;
 }
 
 void timer::reset() {
